feat: 자기이해 Design 결과 화면 구현 (#49) - 화면 재사용을 위해 DesignResultPage 구조 수정
